requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.26.5
idna==3.10
inquirerpy==0.3.4
jinja2==3.1.3
llvmlite==0.43.0
markupsafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
openai-whisper==20240930
packaging==24.2
pfzy==0.3.4
prompt-toolkit==3.0.48
psutil==6.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
sudachidict-full==20241021
sudachipy==0.6.9
sympy==1.13.1
termcolor==2.5.0
tiktoken==0.8.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
tqdm==4.67.1
triton==3.1.0
typing-extensions==4.9.0
urllib3==2.2.3
wcwidth==0.2.13
